job_automation/LinkedinAutomation/ResumeAutomationProMaxUltra++/requirements.txt
pandas
langchain
langsmith
langchain[google-genai]
python-dotenv
openpyxl
odfpy